+++ raspicat_waypoint_navigation/include/raspicat_waypoint_navigation/BaseWaypointParamChange.hpp
