+++ raspicat_waypoint_navigation/include/raspicat_waypoint_navigation/BaseWaypointServer.hpp
