+++ raspicat_waypoint_navigation/include/raspicat_waypoint_navigation/WaypointNavHelperPlugin.hpp
